+++ dwds/test/chrome_proxy_service_test.dart
@@ -854,7 +854,7 @@ void main() {
                     event.isolate.id != initialIsolateId),
               ])));
           service.destroyIsolate();
-          await service.createIsolate();
+          await service.createIsolate(context.appConnection);
           await eventsDone;
           expect((await service.getVM()).isolates.first.id,
               isNot(initialIsolateId));