+++ dwds/test/debugger_test.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 @TestOn('vm')
+import 'dart:async';
 import 'package:dwds/dwds.dart' show ModuleStrategy;
 import 'package:dwds/src/debugging/debugger.dart';
 import 'package:dwds/src/debugging/inspector.dart';
@@ -11,6 +12,8 @@ import 'package:dwds/src/utilities/dart_uri.dart';
 import 'package:dwds/src/utilities/shared.dart';
 import 'package:source_maps/parser.dart';
 import 'package:test/test.dart';
+import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart'
+    show DebuggerPausedEvent;
 
 import 'fixtures/context.dart';
 import 'fixtures/debugger_data.dart';
@@ -20,11 +23,14 @@ final context = TestContext();
 AppInspector inspector;
 Debugger debugger;
 FakeWebkitDebugger webkitDebugger;
+StreamController<DebuggerPausedEvent> pausedController;
 
 void main() async {
   setUpAll(() async {
     globalModuleStrategy = ModuleStrategy.requireJS;
     webkitDebugger = FakeWebkitDebugger();
+    pausedController = StreamController<DebuggerPausedEvent>();
+    webkitDebugger.onPaused = pausedController.stream;
     debugger = await Debugger.create(
       null,
       webkitDebugger,
@@ -64,5 +70,22 @@ void main() async {
     var firstFrame = frames[0];
     var frame1Variables = firstFrame.vars.map((each) => each.name).toList();
     expect(frame1Variables, ['a', 'b']);
+  });
+
+  group('errors', () {
+    setUp(() {
+      // We need to provide an Isolate so that the code doesn't bail out on a null
+      // check before it has a chance to throw.
+      inspector = FakeInspector(fakeIsolate: simpleIsolate);
+    });
+
+    test('errors in the zone are caught and logged', () async {
+      // Add a DebuggerPausedEvent with a null parameter to provoke an error.
+      pausedController.sink.add(DebuggerPausedEvent(null));
+      expect(
+          Debugger.logger.onRecord,
+          emitsThrough(predicate(
+              (log) => log.message == 'Error handling Chrome event')));
+    });
   });
 }