+++ dwds/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.7.5-dev
+
+- Fix an issue where resuming while paused at the start would cause an error.
+
+## 0.7.4
+
+- Deobfuscate DDC extension method stack traces.
+- Properly get all libraries with the `legacy` module system.
+
 ## 0.7.3
 
 - Correctly set `Isolate` state if debugging is initiated after the application